pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tampermint"]

[project]
name = "tampermint"
version = "0.0.1"
authors = [{ name = "Cutipus", email = "Cutipus@proton.me" }]
description = "A simple bot for playing music on Discord."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = ["discord.py[voice]", "PyNaCl", "yt_dlp"]

[project.urls]
Homepage = "https://github.com/cutipus/vigilant-pancake"
Issues = "https://github.com/cutipus/vigilant-pancake/issues"
